requirements.txt
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.3.0
pluggy==0.7.1
py==1.6.0
pylint==2.1.1
pytest==3.8.0
PyYAML==3.13
selenium==3.14.0
six==1.11.0
typed-ast==1.1.0
urllib3==1.23
wrapt==1.10.11
